pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tgrafy-commons"
version = "0.2.10"
description = "Common utilities and interfaces"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
    "boto3>=1.26.0",
    "httpx",
    "PyJWT",
    "requests",
    "cryptography",
    "boto3-stubs",
    "mypy-boto3-ssm",
    "pymongo"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=0.990",
]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true